% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fetch_chain_data}
\alias{fetch_chain_data}
\title{Fetches on-chain data}
\usage{
fetch_chain_data(chain = "Polkadot", api_key)
}
\arguments{
\item{chain}{Network where data are fetched (Polkadot or Kusama).}

\item{api_key}{Key for Subscan API}
}
\value{
A list containing the following objects:
\itemize{
\item \verb{total issuance}: the total token supply
\item \code{nominator_bonded}: the tokens bonded in staking by nominators
\item \code{validator_bonded}: the tokens bonded in staking by validators
\item \code{price}: the token price in USD
}
}
\description{
Fetches on-chain data
}
