% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_issuance}
\alias{get_issuance}
\title{Calculates Token Issuance and Allocations}
\usage{
get_issuance(
  current_supply,
  current_issuance,
  today,
  end_date,
  nodes,
  token_price,
  save = FALSE,
  phase2.opt
)
}
\arguments{
\item{current_supply}{The current token total supply}

\item{current_issuance}{The current token annual issuance}

\item{today}{Today's Date (or any date, keep in mind to adjust \verb{current supply} and \verb{current issuance})}

\item{end_date}{Date to end the fetched data}

\item{nodes}{The total number of nodes producing blocks}

\item{token_price}{The token price in USD}

\item{save}{Option to save the output as a .csv (FALSE by default)}

\item{phase2.opt}{A list of the following objects:
\itemize{
\item \code{phase2_start} A date for the start of Phase II
\item \code{node_reward} rewards in USD per node per month
\item \code{overcollateral_factor} Over-collateralization for minting stablecoins to pay for operations
\item \code{to_resilience} Fraction of the non-reserve to be allocated for resilience
\item \code{to_treasury} Fraction of the non-reserve to be allocated for the Treasury
}}
}
\value{
A dataframe with daily values for:
\itemize{
\item date, year, month and day
\item issuance: the daily issuance of newly minted tokens
\item supply: to total token circulating supply
\item reserve: the daily issuance for reserve
\item non-reserve: the daily issuance not diverted to the reserve
\item to_treasury: the daily non-reserve budget allocated to the treasury
\item operations: the daily non-reserve budget allocated to the operations
\item to_resilience: the daily non-reserve budget allocated to the resilience
\item to_stakers: the daily non-reserve budget allocated to the stakers
}
}
\description{
Calculates Token Issuance and Allocations
}
