% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_summary}
\alias{get_summary}
\title{Calculates Summary Statistics}
\usage{
get_summary(
  data,
  year,
  staking_rate,
  stock_info,
  token_price,
  budget,
  nodes,
  phase2.opt = NULL
)
}
\arguments{
\item{data}{Output from the \code{get_issuance} function}

\item{year}{Year to calculate the summary statistics}

\item{staking_rate}{A vector containing staking rates values}

\item{stock_info}{A list containing the held token supply (\code{held}) and which percentage is bonded in staking (\code{bonded_pct})}

\item{token_price}{A vector containing token price values in USD}

\item{budget}{The requested annual budget}

\item{nodes}{The total number of nodes producing blocks}

\item{phase2.opt}{A list of the following objects:
\itemize{
\item \code{self_stakes} a vector for the distribution of self stakes (length must match \code{nodes})
\item \code{self_stake} average own self stake across multiple nodes
\item \code{pct_own_nodes} percentage of own nodes
\item \code{node_cost} costs per node per month (servers, electricity, etc.)
\item \code{FTE} costs for full time engineers
\item \code{node_reward} rewards in USD per node per month
\item \code{overcollateral_factor} Over-collateralization for minting stablecoins to pay for operations
\item \code{start_date} Start date of Phase II
}

Default is \code{NULL}, meaning it assumes no Phase II.}
}
\description{
Calculates Summary Statistics
}
